% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprecessing.R
\name{assert_cols}
\alias{assert_cols}
\title{Asserts if dataset has required columns}
\usage{
assert_cols(data, req_cols)
}
\arguments{
\item{data}{dataset to check}

\item{req_cols}{chr; vector containing required column names}
}
\description{
Asserts if dataset has required columns
}
