% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{build_svm}
\alias{build_svm}
\title{Support vector machine algorithm}
\usage{
build_svm(data, kernel, gamma, cost)
}
\arguments{
\item{data}{list of data.frames; containing the training and testing sets for the model}

\item{kernel}{chr; the kernel used in training and predicting}

\item{gamma}{float; parameter needed for all kernels except linear (default: 1/(data dimension))}

\item{cost}{float; cost of constraints violation (default: 1)—it is the ‘C’-constant of the regularization term in the Lagrange formulation}
}
\value{
list of parameters created by the SVM model
}
\description{
Support vector machine algorithm
}
