% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{building_models}
\alias{building_models}
\title{Builds models}
\usage{
building_models(
  data,
  gamma = 0.1,
  cost = 0.1,
  ntrees = 50,
  mtry = 5,
  importance = TRUE,
  ...
)
}
\arguments{
\item{data}{list of 3; noises, train, test}

\item{gamma}{float; svm; parameter needed for all kernels except linear (default: 1/(data dimension))}

\item{cost}{float; svm; cost of constraints violation (default: 1)—it is the ‘C’-constant of the regularization term in the Lagrange formulation}

\item{ntrees}{int; RandomForest; number of trees to grow}

\item{mtry}{int; RandomForest; number of variables randomly sampled as candidates at each split}

\item{importance}{bool; RandomForest; should importance of predictors be assessed}

\item{eta}{float; xgboost; control the learning rate; scale the contribution of each tree by a factor of 0 < eta < 1; by default 0.3}

\item{nrounds}{int; xgboost; number of iterations;lower value for eta implies larger value; by default 75}

\item{max_depth}{int; xgboost; maximum depth of a tree; by default 6}

\item{subsample}{float; xgboost; subsample ratio of the training instance; 0.5 prevents overfitting; by default 0.5}

\item{colsample_bytree}{float; xgboost; subsample ratio of columns when constructing each tree; by default 0.5}

\item{eval_metric}{chr; xgboost; evaluation metrics for validation data; by default "error"}

\item{objective}{chr; xgboost; specify the learning task and the corresponding learning objective; by default "binary:logistic"}
}
\value{
data.frame; containing models
}
\description{
Builds models
}
