% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprecessing.R
\name{feature_selection}
\alias{feature_selection}
\title{Removes columns where over x% of instances are: zeros or NA
takes columns with sd value that is more than x}
\usage{
feature_selection(data, zeros_perc = 0.95, na_perc = 0.95, sd_val = 1)
}
\arguments{
\item{data}{list of data.frames; original dataset}

\item{zeros_perc}{float; zeros ratio in attribute (above that value attribute will we removed)}

\item{na_perc}{float; NA ratio in attribute (above that value attribute will we removed)}

\item{sd_val}{float; sd value of attribute (below that value attribute will we removed)}
}
\value{
list of 3 tibbles; all, save and rm
\itemize{
  \item all - original set of attributes
  \item save - attributes to save
  \item rm - attributes to remove
}
}
\description{
Removes columns where over x% of instances are: zeros or NA
takes columns with sd value that is more than x
}
