% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{generate_report}
\alias{generate_report}
\title{Generates html report}
\usage{
generate_report(
  data,
  models,
  to_plot = c("aa_000", "ba_003", "cn_005"),
  output_dir = "~/scania/"
)
}
\arguments{
\item{data}{list of 4; preprocessed, noises, train, test}

\item{models}{data.frame; output models}

\item{to_plot}{chr; a vector of attributes names' to plot}

\item{output_dir}{chr; path to save the report}
}
\description{
Generates html report
}
