% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{plot_noise}
\alias{plot_noise}
\title{Plots noised attribute}
\usage{
plot_noise(data, variable, n = 100)
}
\arguments{
\item{data}{list of data.frames with noise}

\item{variable}{chr attribute name to plot}

\item{n}{int number of points to plot}
}
\value{
plot of noise
}
\description{
Plots noised attribute
}
