% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprecessing.R
\name{preprocessing}
\alias{preprocessing}
\title{Prepares data for the models}
\usage{
preprocessing(train_path, test_path, noise_level = c(0, 0.1, 0.2), ...)
}
\arguments{
\item{train_path}{chr; path to training dataset}

\item{test_path}{chr; path to testing dataset}

\item{noise_level}{float or vector of floats; by default c(0, 0.1, 0.2)}

\item{zeros_perc}{float; zeros ratio in attribute (above that value attribute will we removed); by default 0.95}

\item{na_perc}{float; NA ratio in attribute (above that value attribute will we removed); by default 0.95}

\item{sd_val}{float; sd value of attribute (below that value attribute will we removed); by default 1}
}
\value{
list of 4; preprocessed, noises, train, test
\itemize{
  \item preprocessed - preprocessed data.frames
  \item noises - list of dataframes with noised records
  \item train - training data table
  \item test - testint data frame
}
}
\description{
Prepares data for the models
}
