% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprecessing.R
\name{read_data}
\alias{read_data}
\title{Reads data and modifies columns types}
\usage{
read_data(train_path, test_path)
}
\arguments{
\item{train_path}{chr; path to training dataset}

\item{test_path}{chr; path to test dataset}
}
\value{
list of 2 lists; train and test
\itemize{
  \item train - training dataset
  \item test - testing dataset
}
}
\description{
Reads data and modifies columns types
}
