% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprecessing.R
\name{sampling}
\alias{sampling}
\title{Solves undersampling problem
copies minor class and adds noise to it, major class is sampled to match length of minor class
multiple noise_levels can be set, noise_level = c(0, 0.1, 0.3) will result in original minor class and two copies with added noise (thus 6000 obs, 3000 pos and 3000 neg)}
\usage{
sampling(data_sub_na, noise_level)
}
\arguments{
\item{data_sub_na}{list of 2 data.frames; train and test}

\item{noise_level}{float or vector of floats; by default c(0, 0.1, 0.2)}
}
\value{
list of 3;
\itemize{
  \item noises; list of data.frames containing noised training records
  \item train; noised training set
  \item test; testing set
}
}
\description{
Solves undersampling problem
copies minor class and adds noise to it, major class is sampled to match length of minor class
multiple noise_levels can be set, noise_level = c(0, 0.1, 0.3) will result in original minor class and two copies with added noise (thus 6000 obs, 3000 pos and 3000 neg)
}
